Add key for dn42 gateway
Hi,

the current state of ICVPN/dn42 inter-connection is pretty sad, 
connections from dn42 to Freifunk are very rarely possible. 

My network is pretty well connected to the rest of dn42 with a lot of fast peerings, which is why I want to establish a peering. 
I'm not sure about the naming of my keyfile, as it is a little bit different from the regular configurations. 

So long,
Tobias